# flutter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Memory Vault API",
    "description": "Travel Memory Vault - Personal travel journal with AI-powered search",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/trips/{trip_id}/sessions:start": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Start V2 Session",
        "operationId": "start_v2_session_api_v2_trips__trip_id__sessions_start_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2SessionStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/sessions/{client_session_id}:stop": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Stop V2 Session",
        "operationId": "stop_v2_session_api_v2_trips__trip_id__sessions__client_session_id__stop_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "client_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Session Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2SessionStopRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/publish:start": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Publish Start",
        "operationId": "publish_start_api_v2_trips__trip_id__publish_start_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2PublishStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2PublishStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/publish:media-complete": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Publish Media Complete",
        "operationId": "publish_media_complete_api_v2_trips__trip_id__publish_media_complete_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2PublishMediaCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2PublishMediaCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/publish:payload-chunk": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Publish Payload Chunk",
        "operationId": "publish_payload_chunk_api_v2_trips__trip_id__publish_payload_chunk_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2PublishPayloadChunkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2PublishPayloadChunkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/publish:commit": {
      "post": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Publish Commit",
        "operationId": "publish_commit_api_v2_trips__trip_id__publish_commit_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2PublishCommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2PublishCommitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/timeline": {
      "get": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Get V2 Timeline",
        "operationId": "get_v2_timeline_api_v2_trips__trip_id__timeline_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2TimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/trips/{trip_id}/route": {
      "get": {
        "tags": [
          "Live Tracking V2"
        ],
        "summary": "Get V2 Route",
        "operationId": "get_v2_route_api_v2_trips__trip_id__route_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit_segments",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit Segments"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "V2PublishMediaCompleteRequest": {
        "properties": {
          "publish_token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Publish Token"
          },
          "client_job_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Job Id"
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Schema Version"
          },
          "uploaded_media": {
            "items": {
              "$ref": "#/components/schemas/V2UploadedMediaRef"
            },
            "type": "array",
            "title": "Uploaded Media"
          }
        },
        "type": "object",
        "required": [
          "publish_token",
          "client_job_id",
          "schema_version"
        ],
        "title": "V2PublishMediaCompleteRequest"
      },
      "V2PublishMediaCompleteResponse": {
        "properties": {
          "publish_token": {
            "type": "string",
            "title": "Publish Token"
          },
          "manifest_status": {
            "type": "string",
            "enum": [
              "started",
              "failed_retryable",
              "failed_terminal",
              "idempotency_conflict",
              "committed"
            ],
            "title": "Manifest Status"
          },
          "manifest_phase": {
            "type": "string",
            "enum": [
              "start_received",
              "media_verified",
              "chunks_complete",
              "raw_ingest_completed",
              "projection_compiled",
              "finalized"
            ],
            "title": "Manifest Phase"
          },
          "accepted_media_count": {
            "type": "integer",
            "title": "Accepted Media Count"
          }
        },
        "type": "object",
        "required": [
          "publish_token",
          "manifest_status",
          "manifest_phase",
          "accepted_media_count"
        ],
        "title": "V2PublishMediaCompleteResponse"
      },
...
      "V2PublishStartRequest": {
        "properties": {
          "client_job_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Job Id"
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Schema Version"
          },
          "publish_summary": {
            "$ref": "#/components/schemas/V2PublishSummary"
          },
          "media_manifest": {
            "items": {
              "$ref": "#/components/schemas/V2MediaManifestItem"
            },
            "type": "array",
            "title": "Media Manifest"
          },
          "media_manifest_digest": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Media Manifest Digest"
          }
        },
        "type": "object",
        "required": [
          "client_job_id",
          "schema_version",
          "publish_summary",
          "media_manifest_digest"
        ],
        "title": "V2PublishStartRequest"
      },
      "V2PublishStartResponse": {
        "properties": {
          "publish_token": {
            "type": "string",
            "title": "Publish Token"
          },
          "manifest_status": {
            "type": "string",
            "enum": [
              "started",
              "failed_retryable",
              "failed_terminal",
              "idempotency_conflict",
              "committed"
            ],
            "title": "Manifest Status"
          },
          "manifest_phase": {
            "type": "string",
            "enum": [
              "start_received",
              "media_verified",
              "chunks_complete",
              "raw_ingest_completed",
              "projection_compiled",
              "finalized"
            ],
            "title": "Manifest Phase"
          },
          "accepted_media_count": {
            "type": "integer",
            "title": "Accepted Media Count"
          },
          "upload_targets": {
            "items": {
              "$ref": "#/components/schemas/V2UploadTarget"
            },
            "type": "array",
            "title": "Upload Targets"
          }
        },
        "type": "object",
        "required": [
          "publish_token",
          "manifest_status",
          "manifest_phase",
          "accepted_media_count"
        ],
        "title": "V2PublishStartResponse"
      },
...
      "V2SessionResponse": {
        "properties": {
          "session_server_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Server Id"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trip Id"
          },
          "client_session_id": {
            "type": "string",
            "title": "Client Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "stop_server_pending": {
            "type": "boolean",
            "title": "Stop Server Pending"
          },
          "stop_client_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Client Event Id"
          },
          "seal_version": {
            "type": "integer",
            "title": "Seal Version"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "commit_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Token"
          }
        },
        "type": "object",
        "required": [
          "session_server_id",
          "trip_id",
          "client_session_id",
          "status",
          "started_at",
          "stop_server_pending",
          "seal_version"
        ],
        "title": "V2SessionResponse"
      },
      "V2SessionStartRequest": {
        "properties": {
          "client_session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Session Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "device_context": {
            "type": "object",
            "title": "Device Context"
          }
        },
        "type": "object",
        "required": [
          "client_session_id",
          "started_at"
        ],
        "title": "V2SessionStartRequest"
      },
      "V2SessionStopRequest": {
        "properties": {
          "seal_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Seal Version"
          },
          "stop_client_event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Stop Client Event Id"
          },
          "stopped_at": {
            "type": "string",
            "format": "date-time",
            "title": "Stopped At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "client_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Session Id"
          }
        },
        "type": "object",
        "required": [
          "seal_version",
          "stop_client_event_id",
          "stopped_at"
        ],
        "title": "V2SessionStopRequest"
      },
...
    }
  }
}